Loading icon when deleting; confirm delete

+++ lib/widgets/login_form_widget.dart
@@ -20,6 +20,7 @@ class _LoginFormWidgetState extends State<LoginFormWidget> {
   final screenNameController = TextEditingController();
   final roomIdController = TextEditingController();
   final formKey = GlobalKey<FormState>();
+  var isLoading = false;
 
   Future _onLogin() async {
     if (formKey.currentState!.validate()) {
@@ -33,8 +34,30 @@ class _LoginFormWidgetState extends State<LoginFormWidget> {
 
   Future _onDeleteRoom() async {
     if (formKey.currentState!.validate()) {
-      final database = Database(roomIdController.text);
-      await database.deleteRoom();
+      final dialog = AlertDialog(
+        title: const Text('Delete'),
+        content: const Text('Are you sure you want to permanently delete this room?'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.of(context, rootNavigator: true).pop(),
+            child: const Text('Cancel'),
+          ),
+          TextButton(
+            onPressed: () async {
+              Navigator.of(context, rootNavigator: true).pop();
+              final database = Database(roomIdController.text);
+              setState(() => isLoading = true);
+              await database.deleteRoom();
+              setState(() => isLoading = false);
+            },
+            child: const Text('Confirm'),
+          ),
+        ],
+      );
+      showDialog(
+        context: context,
+        builder: (context) => dialog,
+      );
     }
   }
 
@@ -58,52 +81,56 @@ class _LoginFormWidgetState extends State<LoginFormWidget> {
 
   @override
   Widget build(BuildContext context) {
-    loadSavedLogin();
-    return Container(
-      constraints: const BoxConstraints(maxWidth: 300),
-      child: Form(
-        key: formKey,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            Padding(
-              padding: const EdgeInsets.only(left: 8, right: 8, top: 8),
-              child: TextFormField(
-                autofocus: true,
-                textInputAction: TextInputAction.next,
-                validator: FieldValidators.validateScreenName,
-                controller: screenNameController,
-                decoration: const InputDecoration(
-                    border: OutlineInputBorder(),
-                    hintText: 'Screen name'
+    if (isLoading) {
+      return const CircularProgressIndicator();
+    } else {
+      loadSavedLogin();
+      return Container(
+        constraints: const BoxConstraints(maxWidth: 300),
+        child: Form(
+          key: formKey,
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: [
+              Padding(
+                padding: const EdgeInsets.only(left: 8, right: 8, top: 8),
+                child: TextFormField(
+                  autofocus: true,
+                  textInputAction: TextInputAction.next,
+                  validator: FieldValidators.validateScreenName,
+                  controller: screenNameController,
+                  decoration: const InputDecoration(
+                      border: OutlineInputBorder(),
+                      hintText: 'Screen name'
+                  ),
                 ),
               ),
-            ),
-            Padding(
-              padding: const EdgeInsets.only(left: 8, right: 8, top: 8),
-              child: TextFormField(
-                textInputAction: TextInputAction.done,
-                validator: FieldValidators.validateRoomId,
-                controller: roomIdController,
-                decoration: const InputDecoration(
-                    border: OutlineInputBorder(),
-                    hintText: 'Room id'
+              Padding(
+                padding: const EdgeInsets.only(left: 8, right: 8, top: 8),
+                child: TextFormField(
+                  textInputAction: TextInputAction.done,
+                  validator: FieldValidators.validateRoomId,
+                  controller: roomIdController,
+                  decoration: const InputDecoration(
+                      border: OutlineInputBorder(),
+                      hintText: 'Room id'
+                  ),
+                  onFieldSubmitted: (_) => _onLogin(),
                 ),
-                onFieldSubmitted: (_) => _onLogin(),
               ),
-            ),
-            Padding(
-              padding: const EdgeInsets.only(left: 8, right: 8, top: 8),
-              child: ElevatedButton(
-                onPressed: _onLogin,
-                onLongPress: _onDeleteRoom,
-                child: const Text('Enter'),
-              ),
-            )
-          ],
+              Padding(
+                padding: const EdgeInsets.only(left: 8, right: 8, top: 8),
+                child: ElevatedButton(
+                  onPressed: _onLogin,
+                  onLongPress: _onDeleteRoom,
+                  child: const Text('Enter'),
+                ),
+              )
+            ],
+          ),
         ),
-      ),
-    );
+      );
+    }
   }
 }
